코드 4일차

+++ src/main/java/com/example/beenproject/common/Const.java
@@ -2,5 +2,5 @@ package com.example.beenproject.common;
 
 public interface Const {
     long SUCCESS = 1;
-    int FAIL = 0;
+    long FAIL = 0;
 }
