+++ src/main/java/com/example/beenproject/common/ResVo.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ResVo {
-    private int result;
+    private long result;
 }
